Move OrderBillGenerator to services and delete the no longer useful CarportSchematicGenerator
