+++ src/main/java/app/services/OrderBillGenerator.java
@@ -1,4 +1,4 @@
-package app.util;
+package app.services;
 
 import app.entities.Material;
 import app.entities.OrderBillItem;
