--- src/main/java/app/util/CarportSchematicGenerator.java
@@ -1,7 +1,0 @@
-package app.util;
-
-public class CarportSchematicGenerator {
-
-    public void generateSchematic(int width, int length){
-    } // SPØRG!!!!
-}
